+++ bittorrent-integration-tests/src/test/java/org/johnnei/javatorrent/torrent/DownloadTorrentIT.java
@@ -4,14 +4,14 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.Executors;
 
 import org.johnnei.javatorrent.TorrentClient;
-import org.johnnei.javatorrent.network.socket.TcpSocket;
 import org.johnnei.javatorrent.it.EndToEndDownload;
 import org.johnnei.javatorrent.network.ConnectionDegradation;
+import org.johnnei.javatorrent.network.socket.TcpSocket;
 import org.johnnei.javatorrent.phases.PhaseData;
 import org.johnnei.javatorrent.phases.PhaseRegulator;
 import org.johnnei.javatorrent.test.DummyEntity;
 import org.johnnei.javatorrent.torrent.algos.requests.RateBasedLimiter;
-import org.johnnei.javatorrent.tracker.PeerConnector;
+import org.johnnei.javatorrent.tracker.NioPeerConnector;
 import org.johnnei.javatorrent.tracker.UncappedDistributor;
 
 /**
@@ -27,7 +27,7 @@ public class DownloadTorrentIT extends EndToEndDownload {
                 .build())
             .setDownloadPort(DummyEntity.findAvailableTcpPort())
             .setExecutorService(Executors.newScheduledThreadPool(2))
-            .setPeerConnector(PeerConnector::new)
+            .setPeerConnector(tc -> new NioPeerConnector(tc, 4))
             .setRequestLimiter(new RateBasedLimiter())
             .setPeerDistributor(UncappedDistributor::new)
             .registerTrackerProtocol("stub", (s, torrentClient) -> null)