+++ bittorrent-integration-tests/src/test/java/org/johnnei/javatorrent/ut/metadata/DownloadMetadataIT.java
@@ -35,7 +35,7 @@ import org.johnnei.javatorrent.test.DummyEntity;
 import org.johnnei.javatorrent.torrent.Metadata;
 import org.johnnei.javatorrent.torrent.Torrent;
 import org.johnnei.javatorrent.torrent.algos.requests.RateBasedLimiter;
-import org.johnnei.javatorrent.tracker.PeerConnector;
+import org.johnnei.javatorrent.tracker.NioPeerConnector;
 import org.johnnei.javatorrent.tracker.UncappedDistributor;
 import org.johnnei.javatorrent.utils.StringUtils;
 import org.johnnei.junit.jupiter.Folder;
@@ -161,7 +161,7 @@ public class DownloadMetadataIT {
 				.setDownloadPort(DummyEntity.findAvailableTcpPort())
 				.setExecutorService(Executors.newScheduledThreadPool(2))
 				.setRequestLimiter(new RateBasedLimiter())
-				.setPeerConnector(PeerConnector::new)
+				.setPeerConnector(tc -> new NioPeerConnector(tc, 4))
 				.registerModule(new ExtensionModule.Builder()
 						.registerExtension(new UTMetadataExtension(torrentFileFolder, downloadFolder))
 						.build())