+++ bittorrent-integration-tests/src/test/java/org/johnnei/javatorrent/utp/DownloadTorrentWithUtpIT.java
@@ -12,7 +12,7 @@ import org.johnnei.javatorrent.phases.PhaseData;
 import org.johnnei.javatorrent.phases.PhaseRegulator;
 import org.johnnei.javatorrent.test.DummyEntity;
 import org.johnnei.javatorrent.torrent.algos.requests.RateBasedLimiter;
-import org.johnnei.javatorrent.tracker.PeerConnector;
+import org.johnnei.javatorrent.tracker.NioPeerConnector;
 import org.johnnei.javatorrent.tracker.UncappedDistributor;
 
 /**
@@ -38,7 +38,7 @@ public class DownloadTorrentWithUtpIT extends EndToEndDownload {
 				.setRequestLimiter(new RateBasedLimiter())
 				.setDownloadPort(port)
 				.setExecutorService(Executors.newScheduledThreadPool(2))
-				.setPeerConnector(PeerConnector::new)
+				.setPeerConnector(tc -> new NioPeerConnector(tc, 4))
 				.setPeerDistributor(UncappedDistributor::new)
 				.registerTrackerProtocol("stub", (s, torrentClient) -> null)
 				.setPhaseRegulator(new PhaseRegulator.Builder()